New my events page based on user star actions  .
This works by adding a new volunteeropportunity model to the datastore. That links the stable ID to the last base url we've seen for that ID. It also adds some memcache support so we don't need to hit base that often.

The old 'my events' page is now called /memberactivity. There's a link on the new my events page but that is likely temporary.